+++ src/main/java/com/stock/stock/StockApplication.java
@@ -1,4 +1,4 @@
-package com.stock.stocklist;
+package com.stock.stock;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
