+++ src/main/java/com/stock/stock/entity/Stock.java
@@ -1,4 +1,4 @@
-package com.stock.stocklist.entity;
+package com.stock.stock.entity;
 
 import lombok.EqualsAndHashCode;
