+++ src/test/java/com/stock/stock/StockApplicationTests.java
@@ -1,4 +1,4 @@
-package com.stock.stocklist;
+package com.stock.stock;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
